Move consume() from gen to the utils module.
The next few commits, starting with this one, aim to break up
jsparagus.gen a bit, as it's too long, contains too much stuff, and has
implicit circular dependencies that get in the way of implementing mypy.
